Change types of tensor's size-related members to prevent integer overflow (#1151)
* Change types of tensor's size-related members, so that integer overflow does not happen so easily (see #1148)

* Fix missing typecast

+++ dlib/dnn/cpu_dlib.cpp
@@ -1646,14 +1646,14 @@ namespace dlib
                     for (long r = 0; r < gradient_input.nr(); ++r)
                     {
                         const float y = r*y_scale;
-                        const long top    = static_cast<long>(std::floor(y));
-                        const long bottom = std::min(top+1, grad.nr()-1);
+                        const long long top    = static_cast<long long>(std::floor(y));
+                        const long long bottom = std::min(top+1, grad.nr()-1);
                         const float tb_frac = y - top;
                         for (long c = 0; c < gradient_input.nc(); ++c)
                         {
                             const float x = c*x_scale;
-                            const long left   = static_cast<long>(std::floor(x));
-                            const long right  = std::min(left+1, grad.nc()-1);
+                            const long long left   = static_cast<long long>(std::floor(x));
+                            const long long right  = std::min(left+1, grad.nc()-1);
                             const float lr_frac = x - left;
 
                             const float tmp = gi[r*gradient_input_row_stride+c];
